Remove redundant lines of code, create new musicPlayer function using use-sound library
