+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "use-sound": "^4.0.1"
+  }
+}
